refactor(sdk)!: move EnqueueResult to the queue lib
EnqueueResult now lives in @iii-dev/queue / iii-queue. The root SDK depends
on it and re-exports it from the old path as a deprecated shim.
