+++ docs/docs.json
@@ -142,7 +142,8 @@
                   "next/sdk-reference/node-sdk",
                   "next/sdk-reference/python-sdk",
                   "next/sdk-reference/rust-sdk",
-                  "next/sdk-reference/browser-sdk"
+                  "next/sdk-reference/browser-sdk",
+                  "next/sdk-reference/queue-lib"
                 ]
               }
             ]
